+++ lib/ui.dart
@@ -94,6 +94,7 @@ part 'src/ui/deck_layout_panel.dart';
 part 'src/ui/flow_panel.dart';
 part 'src/ui/html_panel.dart';
 part 'src/ui/lazy_panel.dart';
+part 'src/ui/caption_panel.dart';
 
 part 'src/ui/inline_label.dart';
 part 'src/ui/inline_html.dart';